IDE: warn if another tab opens the file and/or changes it and allow to continue as read-only in the former case.
This is not yet a full 'check if file on disk has changed' as we used to have before some 2018 refactoring, because we only guard against another tab opening the file and not against any changes that may happen on disk for other reasons. However, since we have not had any such thing for almost 3 years, it's good that it's at least partially back..
To fully restore that functionality we need to restore the `mtime`-based workflow that also broke around the same time.

Partly addresses https://github.com/BelaPlatform/Bela/issues/648